青鑫：召唤物的一些东西

+++ Code/Prometheus/Assets/Scripts/Logical/Brick/Brick.cs
@@ -286,7 +286,7 @@ public class Brick : GameItemBase, IEquatable<Brick> {
         }
     }
 
-    int bid;
+    ulong bid;
 
     private void ShowBlock()
     {
